Add files via upload
This code allows me to look at the Stocks I am interested in. I look to further improve this code by making it code that will ask the user for a Stock Ticket and ask for the time frame they want to look at the cumulative returns in this time period.